Fix Bug Whenever Player Loses
Whenever a player would lose, the game would get stuck before the time limit. Therefore, the player would have to go through the loseSequence() again. Updating the `while` statement to check if the gameStarted seems to fix the issue. An `if` statement was also added so that when a player would lose, the winSequence() would not accidentally run.